XW-3939 | share middlewares between fastboot prod server and ember serve
